requirements.txt
langchain_community==0.0.16
langchain==0.1.4
langchain_openai==0.0.5
chromadb==0.4.22
chainlit==1.0.200
Chroma==0.2.0
python-dotenv==1.0.1
mlflow==2.10.0
boto3==1.34.33